OSPFkpsp11

Description

Authors : Gruppe 11: Gabriel Zimmerli, Reto Lehnherr

Source : https:github.comGabriel-Zkpsp

OSPFkpsp: Simulation des OSPF Protokoll in Haskell

Projektarbeit im Modul «Konzepte von Programmiersprachen» bei Edgar Lederer, FHNW

Documentation

type Address = String

IP-Adresse eines Routers

type Interface = String

Das Interface des Routers

type Priority = Int

Priorität des Protokoll

type Dead = Int

Timeout bis Route erneuert werden muss

type RouterId = String

ID des Routers gemäss dem OSPF-Protokoll

type Distance = Int

Die Distanz zwischen zwei Routern

type Metric = Int

Das Gewicht des nach OSPF gewichteten Pfades (Aufsummierung der Distanzen auf Pfad)

type Route = [RouterId]

Liste mit den Hops, welche auf dem Pfad gemacht werden müssen

type ChildNode = (RouterId, Distance)

Beschreibt einen Kind Router bezogen zu seinem Eltern Router

type ShortestPath = (RouterId, Metric, Route)

Beschreibung eines kürzesten Pfad zu einer Route

type NeighboursTableEntry = (Address, Interface, State, RouterId, Priority, Dead)

Beschreibt einen Eintrag in der Nachbarschaftstabelle

type GraphEntry = (RouterId, [ChildNode])

Beschreibt einen Eintrag im Topologie Graphen

type NeighboursTable = [NeighboursTableEntry]

Beschreibt die Nachbarschaftstabelle der Routers, dessen Rolle diese Applikation eingenommen hat

type Graph = [GraphEntry]

Der Graph beschreibt die Netzwerktopologie

data State

Status der Verbindung gemäss OSPF

Constructors

Down

Attempt

Init

TwoWay

ExStart

Exchange

Loading

Full

dataLinkFrameHeaderLength :: Int

Header Länge (in Byte) des Data Link Frames

ipPacketHeaderLength :: Int

Header Länge (in Byte) des IP Packet

ospfPacketHeaderLength :: Int

Header Länge (in Byte) des OSPF Packet

messageTypeOffset :: Int

Offset (in Byte) vom Anfang des OSPF Packet bis zum Message Type

messageTypeLength :: Int

Länge (in Byte) des Message Types

routerIdOffset :: Int

Offset (in Byte) vom Anfang des OSPF Packet bis zur Router ID

routerIdLength :: Int

Länge (in Byte) der Router ID

lsuHeaderLength :: Int

Header Länge (in Byte) des Link State Update Packet

numberOfLsaOffset :: Int

Offset (in Byte) vom Anfang des Link State Update Headers bis zur Anzahl Link State Advertisements

numberOfLsaLength :: Int

Länge (in Byte) der Anzahl Link State Advertisements

lsaHeaderLength :: Int

Header Länge (in Byte) des Link State Advertisement Packet

lsaTypeOffset :: Int

Offset (in Byte) vom Anfang des Link State Advertisement Headers bis zum LSA-Type

lsaTypeLength :: Int

Länge (in Byte) des LSA-Type

lsaLengthOffset :: Int

Offset (in Byte) vom Anfang des Link State Advertisement Headers bis zur Link State Advertisement Länge

lsaLengthLength :: Int

Länge (in Byte) der Link State Advertisement Länge

lsaContentLength :: Int

Content Länge (in Byte) des Link State Advertisement Packet (ohne die einzelnen Links)

numberOfLinksOffset :: Int

Offset (in Byte) vom Anfang des Link State Advertisement Contents bis zu den Anzahl Links

numberOfLinksLength :: Int

Länge (in Byte) der Anzahl Links

linkLength :: Int

Länge (in Bytes) eines Links

childRouterIdOffset :: Int

Offset (in Byte) vom Anfang des Links bis zur ID des Child Routers

childRouterIdLength :: Int

Länge (in Byte) der Child Router ID

metricOffset :: Int

Offset (in Byte) vom Anfang des Links bis zur Metrik

metricLength :: Int

Länge (in Byte) der Metrik

hexToIp :: String -> String

Nimt einen Hex-String entgegen und parst diesen in eine IP-Adresse | Parameter 1: Hex-String (IP Adresse)

hexToInt :: String -> Int

Nimt einen Hex-String entgegen und parst diesen in einen Int | Parameter 1: Hex-String

extractMessageType :: String -> Int

Extrahiert aus einem Link State Update (Hex-String) den Message Type (4 für Link State Update) | Parameter 1: Hex-String (Link State Update)

extractRouterId :: String -> RouterId

Extrahiert aus einem Link State Update (Hex-String) die Router Id des Senders | Parameter 1: Hex-String (Link State Update)

extractNumberOfLsa :: String -> Int

Extrahiert aus einem Link State Update (Hex-String) die Anzahl Link State Advertisements, die in diesem Update enthalten sind | Parameter 1: Hex-String (Link State Update)

extractLsaLengths :: String -> Int -> Int -> [Int] -> [Int]

Extrahiert aus einem Link State Update (Hex-String) die Längen der einzelnen Link State Advertisements | Parameter 1: Hex-String (Link State Update) | Parameter 2: Anzahl Link State Advertisements in diesem Link State Update | Parameter 3: Aktuelles Link State Advertisement (am Anfang 1) | Parameter 4: Akkumulator Liste für LSA-Lengths (Am Anfang leer)

extractLsaTypes :: String -> Int -> Int -> [Int] -> [Int] -> [Int]

Extrahiert aus einem Link State Update (Hex-String) die LSA Types der einzelnen Link State Advertisements (1 für Router-LSA) | Parameter 1: Hex-String (Link State Update) | Parameter 2: Anzahl Link State Advertisements in diesem Link State Update | Parameter 3: Aktuelles Link State Advertisement (am Anfang 1) | Parameter 4: Liste mit allen Längen der einzelnen Link State Advertisements | Parameter 5: Akkumulator Liste für LSA-Types (Am Anfang leer)

extractNumbersOfLinks :: String -> Int -> Int -> [Int] -> [Int] -> [Int] -> [Int]

Extrahiert aus einem Link State Update (Hex-String) die Anzahl Links, die in diesem Link State Advertisement enthalten sind | Parameter 1: Hex-String (Link State Update) | Parameter 2: Anzahl Link State Advertisements in diesem Link State Update | Parameter 3: Aktuelles Link State Advertisement (am Anfang 1) | Parameter 4: Liste mit allen Längen der einzelnen Link State Advertisements | Parameter 5: Liste mit allen LSA-Typen der einzelnen Link State Advertisements | Parameter 6: Akkumulator Liste für NumberOfLinks (Am Anfang leer)

extractChildNodes :: String -> Int -> Int -> [Int] -> [Int] -> [Int] -> Int -> [ChildNode] -> [ChildNode]

Extrahiert aus einem Link State Update (Hex-String) die Router Id und entsprechenden Distanzen der Child Router | Parameter 1: Hex-String (Link State Update) | Parameter 2: Anzahl Link State Advertisements in diesem Link State Update | Parameter 3: Aktuelles Link State Advertisement (am Anfang 1) | Parameter 4: Liste mit allen Längen der einzelnen Link State Advertisements | Parameter 5: Liste mit allen LSA-Typen der einzelnen Link State Advertisements | Parameter 6: Liste mit den Anzahl Links pro Link State Advertisement (nur für Router-LSA) | Parameter 7: Aktueller Link (am Anfang 1) | Parameter 8: Akkumulator Liste für ChildNodes (Am Anfang leer)

readLinkStateUpdate :: FilePath -> IO GraphEntry

Liest ein LinkStateUpdate aus einem File ein und gibt die Topologie Tabelle (Graph) als Liste zurück | Parameter 1: Datei Pfad zum Link State Update File das eingelesen werden soll

buildTopoGraph :: IO [GraphEntry]

Liest alle definierten Link State Updates ein und baut daraus den Topology Graph

setupRoute :: Graph -> [ShortestPath]

Initialisiere die Routes Tabelle anhand des Graphen. Die Metriken für jeden Pfad werden mit -1 initialisiert

nextRouterId :: Graph -> [ShortestPath] -> RouterId

Gibt RouterId zurück, welcher noch nicht abgearbeitet ist und den aktuell kürzesten Pfad hat

sortShortestPath :: (RouterId, Metric) -> ShortestPath -> (RouterId, Metric)

Vergleicht ein (RouterId, Metric)-Tupel mit einem ShortestPath und gibt das Element mit kleinerer Metrik zurück, als (RouterId, Metric) | Wird verwendet um den Router mit dem aktuell kürzesten Pfad zu finden

updateShortestPaths :: [ChildNode] -> ShortestPath -> [ShortestPath] -> [ShortestPath]

Aktualisiert die Liste der kürzesten Pfade anhand des aktuell ausgewählten Router. | Das erste Argument beschreibt die Distanz zu den Nachbarn des Routers | Das zweite Argument ist der kürzeste Pfad zum aktuellen Router | Das letzte Argument ist die Liste der aktuell kürzesten Pfade. | Die Rückgabe ist die aktualisierte Shortestpath-Liste

currentRoute :: [ShortestPath] -> RouterId -> ShortestPath

Gibt den aktuellen ShortestPath zu einem Router aus der Liste zurück

neighboursOfCurrentRouter :: RouterId -> Graph -> [ChildNode]

Finde die Nachbarn im Graph für den aktuellen Router

dijkstra :: Graph -> [ShortestPath] -> [ShortestPath]

Berechung der kürzesten Pfade. Nimmt Topologie Tabelle als input Parameter und gibt eine Liste von Trippeln (ID, Metrik, Route) zurück

printRoutingTable :: [ShortestPath] -> NeighboursTable -> IO ()

Generiert die Routing Tabelle und schreibt diese auf die Konsole

printGraph :: Graph -> IO ()

Schreibt den Graphen auf die Konsole

printShortestPaths :: [ShortestPath] -> IO ()

Schreibt die Shortestpath Trees auf die Konsole

splitStringOnTab :: String -> [String]

Teilt einen String an den Tabulatorenzeichen

parseNeighbourTableLine :: [String] -> NeighboursTableEntry

Parst eine Zeile in der Nachbarschaftstabelle

readNeighbourTable :: String -> NeighboursTable

Liest die Nachbarschaftstabelle (in Tabellenform) aus einem File ein und parst diese in eine Liste mit Nachbarn und deren Eigenschaften

main :: IO ()

Das Programm.